api/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.30
alembic>=1.13.0
asyncpg>=0.29.0
# psycopg2-binary removido - não necessário para async, asyncpg é suficiente

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Validation
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.1.0

# Utils
python-dotenv>=1.0.0

# File uploads
aiofiles>=23.2.0

# CORS
# (included in fastapi)

# Testing (opcional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
